Move SyncDBDataSource to database layer
